+++ packages/macros/package.json
@@ -15,6 +15,7 @@
   },
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.2.0",
+    "@embroider/test-support": "0.0.8",
     "@types/babel__core": "^7.0.4",
     "@types/babel__generator": "^7.0.1",
     "@types/babel__template": "^7.0.1",
@@ -31,7 +32,6 @@
     "@babel/traverse": "^7.2.4",
     "@babel/types": "^7.3.2",
     "@embroider/core": "0.0.8",
-    "babel-core": "^6.26.3",
     "resolve": "^1.8.1",
     "semver": "^5.6.0"
   },
