+++ test-packages/support/package.json
@@ -3,10 +3,14 @@
   "version": "0.0.8",
   "private": true,
   "dependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.2.0",
+    "@babel/core": "^7.2.2",
     "@types/fs-extra": "^5.0.4",
     "@types/node": "^10.5.2",
+    "babel-core": "^6.26.3",
     "console-ui": "^3.0.0",
     "ember-cli": "~3.7.1",
-    "fs-extra": "^7.0.0"
+    "fs-extra": "^7.0.0",
+    "qunit": "^2.8.0"
   }
 }
